Add support for multiple ex file upload and temp angular view

+++ src/main/java/com/moe/metricsconsumer/controllers/RawDataController.java
@@ -4,8 +4,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.moe.metricsconsumer.models.ExerciseDocument;
 
+import com.moe.metricsconsumer.models.measureSummary.MeasureSummary;
 import com.moe.metricsconsumer.repositories.AchievementRepository;
 import com.moe.metricsconsumer.repositories.ExerciseDocumentRepository;
+import com.moe.metricsconsumer.repositories.MeasureRepository;
 import com.moe.metricsconsumer.repositories.XmlRepository;
 import org.bson.BsonBinarySubType;
 import org.bson.types.Binary;
@@ -14,9 +16,13 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.lang.NonNull;
+import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+
+import java.util.ArrayList;
+import java.util.List;
 
 
 /**
@@ -36,6 +42,9 @@ public class RawDataController {
   XmlRepository xmlRepository;
 
   @Autowired
+  private MeasureRepository measureRepository;
+
+  @Autowired
   private AchievementRepository achievementRepository;
 
   @Autowired
@@ -50,7 +59,7 @@ public class RawDataController {
   // Flow of information - see activity diagram -> need to be sent with a code
   @PostMapping("/")
   @ResponseBody
-  public ObjectNode newStudentCode(@RequestHeader(value="measureSummaryRef") String measureSummaryRef,
+  public ObjectNode newStudentCode(@Nullable @RequestHeader(value="measureSummaryRef") String measureSummaryRef,
                                    @RequestParam("uploadingFiles") MultipartFile[] uploadingFiles){
     return saveExFiles(measureSummaryRef, uploadingFiles);
   }
@@ -63,26 +72,42 @@ public class RawDataController {
     return saveExFiles(measureSummaryRef, uploadingFiles);
   }
 
-  private ObjectNode saveExFiles(@NonNull String measureSummaryRef, MultipartFile[] uploadingFiles) {
+  private ObjectNode saveExFiles(String measureSummaryRefParam, MultipartFile[] uploadingFiles) {
+    String measureSummaryRef = measureSummaryRefParam;
+    if (measureSummaryRef == null) {
+      MeasureSummary measureSummary = new MeasureSummary();
+      // TODO: calculate measureSummary from MultipartFile[]
+      // call calculateMeasureSummaryFromExFiles(uploadingFiles, null)
+      MeasureSummary savedMeasureSummary = this.measureRepository.save(measureSummary);
+      measureSummaryRef = savedMeasureSummary.getId();
+    } else {
+      // TODO: recalculate measure summary from MultipartFile[]
+      // call calculateMeasureSummaryFromExFiles(uploadingFiles, measureSummaryRef)
+      try {
+        xmlRepository.removeByMeasureSummaryRef(measureSummaryRef);
+      } catch (Exception e) {
+        logger.error(e.toString());
+      }
+    }
 
-    try {
-      xmlRepository.removeByMeasureSummaryRef(measureSummaryRef);
-    } catch (Exception e) {}
+    logger.debug("measureSummaryRef: " + measureSummaryRef);
 
+    List<ExerciseDocument> exerciseDocumentList = new ArrayList<>();
     for(MultipartFile uploadedFile : uploadingFiles) {
-      // TODO: there might be more than one file!!! fix this for bulk upload
       try {
         ExerciseDocument exerciseDocument = new ExerciseDocument(
           measureSummaryRef,
           "ex",
+          uploadedFile.getOriginalFilename(),
           new Binary(BsonBinarySubType.BINARY, uploadedFile.getBytes())
         );
-        exerciseDocumentRepository.save(exerciseDocument);
+        exerciseDocumentList.add(exerciseDocument);
         logger.debug(exerciseDocument.toString());
       } catch (Exception e) {
         e.printStackTrace();
         return controllerUtil.jsonResponse(4000, "ERROR");
       }
+      exerciseDocumentRepository.saveAll(exerciseDocumentList);
     }
 
     return controllerUtil.jsonResponse(2000, "Success - document has been saved");
